extract: defineContent logic from AddTemplateContext
The goal is to make AddTemplateContext only group
the individual parts that make up a template request,
above the stores for each step that deal with data
important for those respective steps.
All other logic is to be extracted out.
This is the first step that takes out defineContent logic,
together with the hardcoded urls in templateHelpers.
