+++ src/features/createAndEditTemplate/defineContent/store/DefineContentStore.tsx
@@ -1,7 +1,9 @@
 import { useAddTemplateContext } from 'features/createAndEditTemplate/workflow/store/AddTemplateContext';
-import { createContext, ReactNode, useContext, useState } from 'react';
-import { NameLabel } from 'services/Content/ContentApi';
+import { createContext, ReactNode, useContext, useEffect, useState } from 'react';
+import { ContentOrigin, NameLabel } from 'services/Content/ContentApi';
+import { useContentListQuery, useRepositoryParams } from 'services/Content/ContentQueries';
 import { TemplateRequest } from 'services/Templates/TemplateApi';
+import { hardcodeRedHatReposByArchAndVersion } from '../core/templateHelpers';
 
 type DefineContentApiType = {
   distribution_versions: NameLabel[];
@@ -41,17 +43,71 @@ type DefineContentStoreType = {
 export const DefineContentStore = ({ children }: DefineContentStoreType) => {
   const [archOpen, setArchOpen] = useState(false);
   const [versionOpen, setVersionOpen] = useState(false);
-
-  // TODO: move here the code from AddTemplateContext
-  // which is dealing only with this step
+  const [hardcodedRedhatRepositories, setHardcodeRepositories] = useState<string[]>([]);
 
   const {
     isEdit,
     templateRequest,
     setTemplateRequest,
-    distribution_versions,
-    distribution_arches,
+    selectedRedhatRepos,
+    setSelectedCustomRepos,
+    setSelectedRedhatRepos,
+    setHardcodeRepositoryUUIDS,
+    editUUID,
   } = useAddTemplateContext();
+
+  // >>>>>>>>
+  // get archs and versions to populate dropdowns
+  const {
+    data: { distribution_versions, distribution_arches } = {
+      distribution_versions: [],
+      distribution_arches: [],
+    },
+  } = useRepositoryParams();
+  // <<<<<<<<
+
+  // >>>>>>>>
+  // 2. fetch those hardcoded repositories
+  const { data } = useContentListQuery(
+    1,
+    10,
+    { urls: hardcodedRedhatRepositories },
+    '',
+    [ContentOrigin.REDHAT],
+    !!hardcodedRedhatRepositories.length,
+  );
+
+  // 1. when a user selects arch and os
+  // get urls of harcoded repos
+  useEffect(() => {
+    if (!!templateRequest.arch && !!templateRequest.version) {
+      const result = hardcodeRedHatReposByArchAndVersion(
+        templateRequest.arch,
+        templateRequest.version,
+      );
+      if (result) {
+        setHardcodeRepositories(result);
+      }
+      if (!editUUID) setSelectedCustomRepos(new Set());
+    }
+  }, [templateRequest.version, templateRequest.arch, editUUID]);
+
+  // 3. filter out hardcoded uuids
+  useEffect(() => {
+    if (data?.data?.length) {
+      const hardcodedItems = data?.data.map((item) => item.uuid);
+
+      setHardcodeRepositoryUUIDS(new Set(hardcodedItems));
+      setSelectedRedhatRepos(
+        new Set(
+          selectedRedhatRepos.has(hardcodedItems[0])
+            ? [...selectedRedhatRepos, ...hardcodedItems]
+            : hardcodedItems,
+        ),
+      );
+    }
+  }, [data?.data]);
+  // <<<<<<<<
 
   const archesDisplay = (arch?: string) =>
     distribution_arches.find(({ label }) => arch === label)?.name || 'Select architecture';
