+++ src/features/createAndEditTemplate/workflow/store/AddTemplateContext.tsx
@@ -9,9 +9,8 @@ import {
 } from 'react';
 import { TemplateRequest } from 'services/Templates/TemplateApi';
 import { QueryClient, useQueryClient } from 'react-query';
-import { useContentListQuery, useRepositoryParams } from 'services/Content/ContentQueries';
-import { ContentOrigin, NameLabel } from 'services/Content/ContentApi';
-import { hardcodeRedHatReposByArchAndVersion } from '../core/templateHelpers';
+import { useContentListQuery } from 'services/Content/ContentQueries';
+import { ContentOrigin } from 'services/Content/ContentApi';
 import { useNavigate } from 'react-router-dom';
 import { useFetchTemplate } from 'services/Templates/TemplateQueries';
 import useRootPath from 'Hooks/useRootPath';
@@ -20,8 +19,6 @@ import useSafeUUIDParam from 'Hooks/useSafeUUIDParam';
 
 export interface AddTemplateContextInterface {
   queryClient: QueryClient;
-  distribution_arches: NameLabel[];
-  distribution_versions: NameLabel[];
   templateRequest: Partial<TemplateRequest>;
   setTemplateRequest: (value: React.SetStateAction<Partial<TemplateRequest>>) => void;
   selectedRedhatRepos: Set<string>;
@@ -29,6 +26,7 @@ export interface AddTemplateContextInterface {
   selectedCustomRepos: Set<string>;
   setSelectedCustomRepos: (uuidSet: Set<string>) => void;
   hardcodedRedhatRepositoryUUIDS: Set<string>;
+  setHardcodeRepositoryUUIDS: (uuidSet: Set<string>) => void;
   checkIfCurrentStepValid: (index: number) => boolean;
   isEdit?: boolean;
   editUUID?: string;
@@ -38,16 +36,16 @@ export const AddTemplateContext = createContext({} as AddTemplateContextInterfac
 
 export const AddTemplateContextProvider = ({ children }: { children: ReactNode }) => {
   const uuid = useSafeUUIDParam('templateUUID');
-  const { data: editTemplateData, isError } = useFetchTemplate(uuid, !!uuid);
+  const { data: editTemplateData, isError } = useFetchTemplate(uuid!, !!uuid);
 
   const navigate = useNavigate();
   const rootPath = useRootPath();
 
   if (isError) navigate(rootPath);
+
   const [templateRequest, setTemplateRequest] = useState<Partial<TemplateRequest>>({});
   const [selectedRedhatRepos, setSelectedRedhatRepos] = useState<Set<string>>(new Set());
   const [selectedCustomRepos, setSelectedCustomRepos] = useState<Set<string>>(new Set());
-  const [hardcodedRedhatRepositories, setHardcodeRepositories] = useState<string[]>([]);
   const [hardcodedRedhatRepositoryUUIDS, setHardcodeRepositoryUUIDS] = useState<Set<string>>(
     new Set(),
   );
@@ -75,15 +73,6 @@ export const AddTemplateContextProvider = ({ children }: { children: ReactNode }
 
   const queryClient = useQueryClient();
 
-  const { data } = useContentListQuery(
-    1,
-    10,
-    { urls: hardcodedRedhatRepositories },
-    '',
-    [ContentOrigin.REDHAT],
-    !!hardcodedRedhatRepositories.length,
-  );
-
   const { data: existingRepositoryInformation, isLoading } = useContentListQuery(
     1,
     10,
@@ -92,34 +81,6 @@ export const AddTemplateContextProvider = ({ children }: { children: ReactNode }
     [ContentOrigin.ALL],
     !!uuid && !!editTemplateData?.repository_uuids.length,
   );
-
-  useEffect(() => {
-    if (!!templateRequest.arch && !!templateRequest.version) {
-      const result = hardcodeRedHatReposByArchAndVersion(
-        templateRequest.arch,
-        templateRequest.version,
-      );
-      if (result) {
-        setHardcodeRepositories(result);
-      }
-      if (!uuid) setSelectedCustomRepos(new Set());
-    }
-  }, [templateRequest.version, templateRequest.arch, uuid]);
-
-  useEffect(() => {
-    if (data?.data?.length) {
-      const hardcodedItems = data?.data.map((item) => item.uuid);
-
-      setHardcodeRepositoryUUIDS(new Set(hardcodedItems));
-      setSelectedRedhatRepos(
-        new Set(
-          selectedRedhatRepos.has(hardcodedItems[0])
-            ? [...selectedRedhatRepos, ...hardcodedItems]
-            : hardcodedItems,
-        ),
-      );
-    }
-  }, [data?.data]);
 
   // If editing, we want to load in the current data
   useEffect(() => {
@@ -162,26 +123,18 @@ export const AddTemplateContextProvider = ({ children }: { children: ReactNode }
     }));
   }, [templateRequestDependencies]);
 
-  const {
-    data: { distribution_versions, distribution_arches } = {
-      distribution_versions: [],
-      distribution_arches: [],
-    },
-  } = useRepositoryParams();
-
   return (
     <AddTemplateContext.Provider
       key={uuid}
       value={{
         queryClient,
-        distribution_arches,
-        distribution_versions,
         templateRequest,
         setTemplateRequest,
         selectedRedhatRepos,
         setSelectedRedhatRepos,
         selectedCustomRepos,
         setSelectedCustomRepos,
+        setHardcodeRepositoryUUIDS,
         hardcodedRedhatRepositoryUUIDS,
         checkIfCurrentStepValid,
         isEdit: !!uuid,
